Add journal tracking for tag change
This feature is extremely hackish, because it is based on the existing
fields from the Journal and JournalDetails model of redmine. By hackish,
I mean that the column were manipulated to store data that are not so
obvious, as a reference, here are the meaning of the columns when the
property of the `journals` table is a `tag`.

Add a new tag

If a new tag was added, then the prop_key will point to a
tag_descriptor, old value will be null, and the new value will be the id
of the new tag

Update a tag
Prop key will be the new tag id, old value is severtiy and value is null

Remove a tag
Pop key will be the old tag id, old value is the verity and value is the
removed value

Any change can still happen, and the latest change will be included
inside of the code as documentation