Implement tagging and tweak css
=== CSS ===
* The CSS was tweaked so that multiple tag row would have equal space in
  between them.
* Also the float right was removed as uusually there is a
  next/previous button to navigate between the issues.

=== The Journal ===
BUG: When there are concurreny editing on the issue, the journal will
not work, but the tag will be saved correctly.

The journal was hacked in order to implement the history. The reason of
hacking is because I did not wanted to add an extra column to the
journal tables. So the existing rows are used. The journal must be able
to construct and to store enough information so that it can build the
tag history without references to tags, because tags are deleted and
recreated everytime. Only referneces it can contains are probably the
tag descriptor ID even then, when a tag descriptor is removed, the
history will be lost. Luckly, the journal tables contains enough columns
to describe the history. Here is the breakdown of how the rows are used
in order to build the tag history.

Prop Key   | old_value    | value
-------------------------------------------
    X      |     nil      | X                  ADD Tag
    X      |     X        | nil                REMOVE Tag
    X      |     X        | X                  Update Tag

Add Tag: Prop store severity, value store description
Remove Tag: Prop store severity, old_value store description
Update Tag: Prop store description, old_value store old severity, value
            store new severtiy